+++ GeosCore/planeflight_mod.F90
@@ -2174,16 +2174,16 @@ CONTAINS
              ! ISORROPIA/HETP H+ and pH (eam, 06/2015)
              !---------------------------------------------------------------
              CASE( 5001 )
-                VARI(V) = State_Chm%IsorropHplus(I,J,L,1)
+                VARI(V) = State_Chm%AteHplus(I,J,L,1)
 
              CASE( 5002 )
-                VARI(V) = State_Chm%IsorropAeropH(I,J,L,1)
+                VARI(V) = State_Chm%AteAeropH(I,J,L,1)
 
              CASE( 5003 )
-                VARI(V) = State_Chm%IsorropAeroH2O(I,J,L,1)
+                VARI(V) = State_Chm%AteAeroH2O(I,J,L,1)
 
              CASE( 5004 )
-                VARI(V) = State_Chm%IsorropBisulfate(I,J,L)
+                VARI(V) = State_Chm%AteBisulfate(I,J,L)
 
              !---------------------------------------------------------------
              ! Local Time (eam, 06/2015)