+++ Headers/state_chm_mod.F90
@@ -21,7 +21,7 @@ MODULE State_Chm_Mod
 !
 ! USES:
 !
-  USE AerMass_Container_Mod              ! Aerosol mass object          
+  USE AerMass_Container_Mod              ! Aerosol mass object
   USE Dictionary_M, ONLY : dictionary_t  ! Fortran hash table type
   USE ErrCode_Mod                        ! Error handling
   USE Phot_Container_Mod                 ! For photolysis state object
@@ -194,15 +194,15 @@ MODULE State_Chm_Mod
      !-----------------------------------------------------------------------
      ! For isoprene SOA via ISORROPIA/HETP
      !-----------------------------------------------------------------------
-     REAL(fp),          POINTER :: IsorropAeropH  (:,:,:,:) ! ISORROPIA aero pH
-     REAL(fp),          POINTER :: IsorropHplus   (:,:,:,:) ! H+ conc [M]
-     REAL(fp),          POINTER :: IsorropIONIC   (:,:,:) ! ISORROPIA aero IONIC
-     REAL(fp),          POINTER :: IsorropOH      (:,:,:) ! ISORROPIA aero OH
-     REAL(fp),          POINTER :: IsorropAeroH2O (:,:,:,:) ! ISORROPIA aero H2O
-     REAL(fp),          POINTER :: IsorropSulfate (:,:,:  ) ! Sulfate conc [M]
-     REAL(fp),          POINTER :: IsorropNitrate (:,:,:,:) ! Nitrate conc [M]
-     REAL(fp),          POINTER :: IsorropChloride(:,:,:,:) ! Chloride conc [M]
-     REAL(fp),          POINTER :: IsorropBisulfate(:,:,:  )! Bisulfate conc [M]
+     REAL(fp),          POINTER :: AteAeropH  (:,:,:,:) ! ISORROPIA aero pH
+     REAL(fp),          POINTER :: AteHplus   (:,:,:,:) ! H+ conc [M]
+     REAL(fp),          POINTER :: AteIONIC   (:,:,:) ! ISORROPIA aero IONIC
+     REAL(fp),          POINTER :: AteOH      (:,:,:) ! ISORROPIA aero OH
+     REAL(fp),          POINTER :: AteAeroH2O (:,:,:,:) ! ISORROPIA aero H2O
+     REAL(fp),          POINTER :: AteSulfate (:,:,:  ) ! Sulfate conc [M]
+     REAL(fp),          POINTER :: AteNitrate (:,:,:,:) ! Nitrate conc [M]
+     REAL(fp),          POINTER :: AteChloride(:,:,:,:) ! Chloride conc [M]
+     REAL(fp),          POINTER :: AteBisulfate(:,:,:  )! Bisulfate conc [M]
 
      !-----------------------------------------------------------------------
      ! For the tagged Hg simulation
@@ -367,11 +367,11 @@ MODULE State_Chm_Mod
      TYPE(dictionary_t)         :: RegDict              ! Registry lookup table
 
      !-----------------------------------------------------------------------
-     ! GEOS specific fields 
+     ! GEOS specific fields
      !-----------------------------------------------------------------------
 #if defined( MODEL_GEOS )
      ! CO mesosphere boundary
-     INTEGER            :: COmesosphere 
+     INTEGER            :: COmesosphere
      CHARACTER(LEN=255) :: impCOmeso
      ! CO2 photolysis
      INTEGER            :: CO2fromGOCART
@@ -607,15 +607,15 @@ CONTAINS
     State_Chm%QQRain            => NULL()
 
     ! Isoprene SOA
-    State_Chm%IsorropAeropH     => NULL()
-    State_Chm%IsorropHplus      => NULL()
-    State_Chm%IsorropAeroH2O    => NULL()
-    State_Chm%IsorropIONIC      => NULL()
-    State_Chm%IsorropOH         => NULL()
-    State_Chm%IsorropSulfate    => NULL()
-    State_Chm%IsorropNitrate    => NULL()
-    State_Chm%IsorropChloride   => NULL()
-    State_Chm%IsorropBisulfate  => NULL()
+    State_Chm%AteAeropH     => NULL()
+    State_Chm%AteHplus      => NULL()
+    State_Chm%AteAeroH2O    => NULL()
+    State_Chm%AteIONIC      => NULL()
+    State_Chm%AteOH         => NULL()
+    State_Chm%AteSulfate    => NULL()
+    State_Chm%AteNitrate    => NULL()
+    State_Chm%AteChloride   => NULL()
+    State_Chm%AteBisulfate  => NULL()
 
     ! Hg simulation quantities
     State_Chm%OceanHg0          => NULL()
@@ -640,9 +640,9 @@ CONTAINS
 
 #if defined( MODEL_GEOS )
     State_Chm%COmesosphere      = .FALSE.
-    State_Chm%impCOmeso         = "unknown" 
+    State_Chm%impCOmeso         = "unknown"
     State_Chm%CO2fromGOCART     = .FALSE.
-    State_Chm%impCO2name        = "unknown" 
+    State_Chm%impCO2name        = "unknown"
     State_Chm%numphoto          = 0
     State_Chm%nxdo              = 0
     State_Chm%nlam              = 0
@@ -1068,7 +1068,7 @@ CONTAINS
           CALL GC_Error( errMsg, RC, thisLoc )
           RETURN
        ENDIF
-       
+
        !---------------------------------------------------------------------
        ! AeroArea
        !---------------------------------------------------------------------
@@ -1288,7 +1288,7 @@ CONTAINS
        !---------------------------------------------------------------------
        ! Gamma SO2
        !---------------------------------------------------------------------
-       
+
        fieldId(1) = 'GammaSO2TMI'
        fieldId(2) = 'GammaSO2O3'
        fieldId(3) = 'GammaSO2H2O2'
@@ -1304,7 +1304,7 @@ CONTAINS
                nSlots     =  5,                                                  &
                nCat       = N,                                                   &
                RC         = RC                                                 )
-          
+
           IF ( RC /= GC_SUCCESS ) THEN
              errMsg = TRIM( errMsg_ir ) // TRIM( chmId )
              CALL GC_Error( errMsg, RC, thisLoc )
@@ -1386,10 +1386,10 @@ CONTAINS
        ENDIF
 
        !---------------------------------------------------------------------
-       ! IsorropAeropH
+       ! AteAeropH
        !---------------------------------------------------------------------
-       fieldId(1) = 'IsorropAeropHAccum'
-       fieldId(2) = 'IsorropAeropHCoarse'
+       fieldId(1) = 'AteAeropHAccum'
+       fieldId(2) = 'AteAeropHCoarse'
 
        ! Allocate and register each field individually
        DO N = 1, 2
@@ -1398,7 +1398,7 @@ CONTAINS
                State_Chm  = State_Chm,                                       &
                State_Grid = State_Grid,                                      &
                chmId      = TRIM( fieldId(N) ),                              &
-               Ptr2Data   = State_Chm%IsorropAeropH,                         &
+               Ptr2Data   = State_Chm%AteAeropH,                         &
                nSlots     = 2,                                               &
                nCat       = N,                                               &
                RC         = RC                                              )
@@ -1411,10 +1411,10 @@ CONTAINS
        ENDDO
 
        !---------------------------------------------------------------------
-       ! IsorropHplus
+       ! AteHplus
        !---------------------------------------------------------------------
-       fieldId(1) = 'IsorropHplusAccum'
-       fieldId(2) = 'IsorropHplusCoarse'
+       fieldId(1) = 'AteHplusAccum'
+       fieldId(2) = 'AteHplusCoarse'
 
        ! Allocate and register each field individually
        DO N = 1, 2
@@ -1423,7 +1423,7 @@ CONTAINS
                State_Chm  = State_Chm,                                       &
                State_Grid = State_Grid,                                      &
                chmId      = TRIM( fieldId(N) ),                              &
-               Ptr2Data   = State_Chm%IsorropHplus,                          &
+               Ptr2Data   = State_Chm%AteHplus,                          &
                nSlots     = 2,                                               &
                nCat       = N,                                               &
                RC         = RC                                              )
@@ -1436,10 +1436,10 @@ CONTAINS
        ENDDO
 
        !---------------------------------------------------------------------
-       ! IsorropAeroH2O
+       ! AteAeroH2O
        !---------------------------------------------------------------------
-       fieldId(1) = 'IsorropAeroH2OAccum'
-       fieldId(2) = 'IsorropAeroH2OCoarse'
+       fieldId(1) = 'AteAeroH2OAccum'
+       fieldId(2) = 'AteAeroH2OCoarse'
 
        ! Allocate and register each field individually
        DO N = 1, 2
@@ -1448,7 +1448,7 @@ CONTAINS
                State_Chm  = State_Chm,                                       &
                State_Grid = State_Grid,                                      &
                chmId      = TRIM( fieldId(N) ),                              &
-               Ptr2Data   = State_Chm%IsorropAeroH2O,                        &
+               Ptr2Data   = State_Chm%AteAeroH2O,                        &
                nSlots     = 2,                                               &
                nCat       = N,                                               &
                RC         = RC                                              )
@@ -1461,15 +1461,15 @@ CONTAINS
        ENDDO
 
        !---------------------------------------------------------------------
-       ! IsorropIONIC
+       ! AteIONIC
        !---------------------------------------------------------------------
-       chmId = 'IsorropIONIC'
+       chmId = 'AteIONIC'
        CALL Init_and_Register(                                               &
             Input_Opt  = Input_Opt,                                          &
             State_Chm  = State_Chm,                                          &
             State_Grid = State_Grid,                                         &
             chmId      = chmId,                                              &
-            Ptr2Data   = State_Chm%IsorropIONIC,                           &
+            Ptr2Data   = State_Chm%AteIONIC,                           &
             RC         = RC                                                 )
 
        IF ( RC /= GC_SUCCESS ) THEN
@@ -1478,15 +1478,15 @@ CONTAINS
           RETURN
        ENDIF
        !---------------------------------------------------------------------
-       ! IsorropOH
+       ! AteOH
        !---------------------------------------------------------------------
-       chmId = 'IsorropOH'
+       chmId = 'AteOH'
        CALL Init_and_Register(                                               &
             Input_Opt  = Input_Opt,                                          &
             State_Chm  = State_Chm,                                          &
             State_Grid = State_Grid,                                         &
             chmId      = chmId,                                              &
-            Ptr2Data   = State_Chm%IsorropOH,                           &
+            Ptr2Data   = State_Chm%AteOH,                           &
             RC         = RC                                                 )
 
        IF ( RC /= GC_SUCCESS ) THEN
@@ -1495,15 +1495,15 @@ CONTAINS
           RETURN
        ENDIF
        !---------------------------------------------------------------------
-       ! IsorropSulfate
+       ! AteSulfate
        !---------------------------------------------------------------------
-       chmId = 'IsorropSulfate'
+       chmId = 'AteSulfate'
        CALL Init_and_Register(                                               &
             Input_Opt  = Input_Opt,                                          &
             State_Chm  = State_Chm,                                          &
             State_Grid = State_Grid,                                         &
             chmId      = chmId,                                              &
-            Ptr2Data   = State_Chm%IsorropSulfate,                           &
+            Ptr2Data   = State_Chm%AteSulfate,                           &
             RC         = RC                                                 )
 
        IF ( RC /= GC_SUCCESS ) THEN
@@ -1514,10 +1514,10 @@ CONTAINS
 
 
        !---------------------------------------------------------------------
-       ! IsorropNitrate
+       ! AteNitrate
        !---------------------------------------------------------------------
-       fieldId(1) = 'IsorropNitrateAccum'
-       fieldId(2) = 'IsorropNitrateCoarse'
+       fieldId(1) = 'AteNitrateAccum'
+       fieldId(2) = 'AteNitrateCoarse'
 
        ! Allocate and register each field individually
        DO N = 1, 2
@@ -1526,7 +1526,7 @@ CONTAINS
                State_Chm  = State_Chm,                                       &
                State_Grid = State_Grid,                                      &
                chmId      = TRIM( fieldId(N) ),                              &
-               Ptr2Data   = State_Chm%IsorropNitrate,                        &
+               Ptr2Data   = State_Chm%AteNitrate,                        &
                nSlots     = 2,                                               &
                nCat       = N,                                               &
                RC         = RC                                              )
@@ -1539,10 +1539,10 @@ CONTAINS
        ENDDO
 
        !---------------------------------------------------------------------
-       ! IsorropChloride
+       ! AteChloride
        !---------------------------------------------------------------------
-       fieldId(1) = 'IsorropChlorideAccum'
-       fieldId(2) = 'IsorropChlorideCoarse'
+       fieldId(1) = 'AteChlorideAccum'
+       fieldId(2) = 'AteChlorideCoarse'
 
        ! Allocate and register each field individually
        DO N = 1, 2
@@ -1551,7 +1551,7 @@ CONTAINS
                State_Chm  = State_Chm,                                       &
                State_Grid = State_Grid,                                      &
                chmId      = TRIM( fieldId(N) ),                              &
-               Ptr2Data   = State_Chm%IsorropChloride,                       &
+               Ptr2Data   = State_Chm%AteChloride,                       &
                nSlots     = 2,                                               &
                nCat       = N,                                               &
                RC         = RC                                              )
@@ -1564,15 +1564,15 @@ CONTAINS
        ENDDO
 
        !---------------------------------------------------------------------
-       ! IsorropBisulfate
+       ! AteBisulfate
        !---------------------------------------------------------------------
-       chmId  = 'IsorropBisulfate'
+       chmId  = 'AteBisulfate'
        CALL Init_and_Register(                                               &
             Input_Opt  = Input_Opt,                                          &
             State_Chm  = State_Chm,                                          &
             State_Grid = State_Grid,                                         &
             chmId      = chmId,                                              &
-            Ptr2Data   = State_Chm%IsorropBisulfate,                         &
+            Ptr2Data   = State_Chm%AteBisulfate,                         &
             RC         = RC                                                 )
 
        IF ( RC /= GC_SUCCESS ) THEN
@@ -1765,7 +1765,7 @@ CONTAINS
           CALL GC_Error( errMsg, RC, thisLoc )
           RETURN
        ENDIF
-       
+
        !---------------------------------------------------------------------
        ! MnII_A
        !---------------------------------------------------------------------
@@ -1800,7 +1800,7 @@ CONTAINS
           CALL GC_Error( errMsg, RC, thisLoc )
           RETURN
        ENDIF
-       
+
        !---------------------------------------------------------------------
        ! MnII_A
        !---------------------------------------------------------------------
@@ -1818,7 +1818,7 @@ CONTAINS
           CALL GC_Error( errMsg, RC, thisLoc )
           RETURN
        ENDIF
-       
+
        !---------------------------------------------------------------------
        ! fupdateHOBr
        !---------------------------------------------------------------------
@@ -2132,7 +2132,7 @@ CONTAINS
        CALL GC_Error( errMsg, RC, thisLoc )
        RETURN
     ENDIF
-    
+
     !------------------------------------------------------------------------
     ! phCloud
     !------------------------------------------------------------------------
@@ -2566,7 +2566,7 @@ CONTAINS
        WRITE( 6, 10 )
  10    FORMAT( /, 'Registered variables contained within the State_Chm object:')
        WRITE( 6, '(a)' ) REPEAT( '=', 79 )
-       
+
        ! Print registered fields
        CALL Registry_Print( Input_Opt   = Input_Opt,                         &
                             Registry    = State_Chm%Registry,                &
@@ -2726,7 +2726,7 @@ CONTAINS
        IF ( RC /= GC_SUCCESS ) RETURN
        State_Chm%Map_KppVar = 0
 
-       ! 
+       !
        ALLOCATE( State_Chm%KPP_AbsTol( N ), STAT=RC )
        CALL GC_CheckVar( 'State_Chm%Kpp_AbsTol', 0, RC )
        IF ( RC /= GC_SUCCESS ) RETURN
@@ -3561,61 +3561,60 @@ CONTAINS
        State_Chm%OMOC_OPOA => NULL()
     ENDIF
 
-    IF ( ASSOCIATED( State_Chm%IsorropAeropH ) ) THEN
-       DEALLOCATE( State_Chm%IsorropAeropH, STAT=RC  )
-       CALL GC_CheckVar( 'State_Chm%IsorropAeropH', 2, RC )
+    IF ( ASSOCIATED( State_Chm%AteAeropH ) ) THEN
+       DEALLOCATE( State_Chm%AteAeropH, STAT=RC  )
+       CALL GC_CheckVar( 'State_Chm%AteAeropH', 2, RC )
        IF ( RC /= GC_SUCCESS ) RETURN
-       State_Chm%IsorropAeropH => NULL()
+       State_Chm%AteAeropH => NULL()
     ENDIF
 
-    IF ( ASSOCIATED( State_Chm%IsorropHplus ) ) THEN
-       DEALLOCATE( State_Chm%IsorropHplus, STAT=RC )
-       CALL GC_CheckVar( 'State_Chm%IsorropHplus', 2, RC )
+    IF ( ASSOCIATED( State_Chm%AteHplus ) ) THEN
+       DEALLOCATE( State_Chm%AteHplus, STAT=RC )
+       CALL GC_CheckVar( 'State_Chm%AteHplus', 2, RC )
        IF ( RC /= GC_SUCCESS ) RETURN
-       State_Chm%IsorropHplus => NULL()
+       State_Chm%AteHplus => NULL()
     ENDIF
 
-    IF ( ASSOCIATED( State_Chm%IsorropOH ) ) THEN
-       DEALLOCATE( State_Chm%IsorropOH, STAT=RC )
-       CALL GC_CheckVar( 'State_Chm%IsorropOH', 2, RC )
+    IF ( ASSOCIATED( State_Chm%AteOH ) ) THEN
+       DEALLOCATE( State_Chm%AteOH, STAT=RC )
+       CALL GC_CheckVar( 'State_Chm%AteOH', 2, RC )
        IF ( RC /= GC_SUCCESS ) RETURN
-       State_Chm%IsorropOH => NULL()
+       State_Chm%AteOH => NULL()
     ENDIF
 
-
-    IF ( ASSOCIATED( State_Chm%IsorropAeroH2O ) ) THEN
-       DEALLOCATE( State_Chm%IsorropAeroH2O, STAT=RC )
-       CALL GC_CheckVar( 'State_Chm%IsorropAeroH2O', 2, RC )
+    IF ( ASSOCIATED( State_Chm%AteAeroH2O ) ) THEN
+       DEALLOCATE( State_Chm%AteAeroH2O, STAT=RC )
+       CALL GC_CheckVar( 'State_Chm%AteAeroH2O', 2, RC )
        IF ( RC /= GC_SUCCESS ) RETURN
-       State_Chm%IsorropAeroH2O => NULL()
+       State_Chm%AteAeroH2O => NULL()
     ENDIF
 
-    IF ( ASSOCIATED( State_Chm%IsorropSulfate ) ) THEN
-       DEALLOCATE( State_Chm%IsorropSulfate, STAT=RC )
-       CALL GC_CheckVar( 'State_Chm%IsorropSulfate', 2, RC )
+    IF ( ASSOCIATED( State_Chm%AteSulfate ) ) THEN
+       DEALLOCATE( State_Chm%AteSulfate, STAT=RC )
+       CALL GC_CheckVar( 'State_Chm%AteSulfate', 2, RC )
        IF ( RC /= GC_SUCCESS ) RETURN
-       State_Chm%IsorropSulfate => NULL()
+       State_Chm%AteSulfate => NULL()
     ENDIF
 
-    IF ( ASSOCIATED( State_Chm%IsorropNitrate ) ) THEN
-       DEALLOCATE( State_Chm%IsorropNitrate, STAT=RC )
-       CALL GC_CheckVar( 'State_Chm%IsorropNitrate', 2, RC )
+    IF ( ASSOCIATED( State_Chm%AteNitrate ) ) THEN
+       DEALLOCATE( State_Chm%AteNitrate, STAT=RC )
+       CALL GC_CheckVar( 'State_Chm%AteNitrate', 2, RC )
        IF ( RC /= GC_SUCCESS ) RETURN
-       State_Chm%IsorropNitrate => NULL()
+       State_Chm%AteNitrate => NULL()
     ENDIF
 
-    IF ( ASSOCIATED( State_Chm%IsorropChloride ) ) THEN
-       DEALLOCATE( State_Chm%IsorropChloride, STAT=RC )
-       CALL GC_CheckVar( 'State_Chm%IsorropChloride', 2, RC )
+    IF ( ASSOCIATED( State_Chm%AteChloride ) ) THEN
+       DEALLOCATE( State_Chm%AteChloride, STAT=RC )
+       CALL GC_CheckVar( 'State_Chm%AteChloride', 2, RC )
        IF ( RC /= GC_SUCCESS ) RETURN
-       State_Chm%IsorropChloride => NULL()
+       State_Chm%AteChloride => NULL()
     ENDIF
 
-    IF ( ASSOCIATED( State_Chm%IsorropBisulfate ) ) THEN
-       DEALLOCATE( State_Chm%IsorropBisulfate, STAT=RC )
-       CALL GC_CheckVar( 'State_Chm%IsorropBisulfate', 2, RC )
+    IF ( ASSOCIATED( State_Chm%AteBisulfate ) ) THEN
+       DEALLOCATE( State_Chm%AteBisulfate, STAT=RC )
+       CALL GC_CheckVar( 'State_Chm%AteBisulfate', 2, RC )
        IF ( RC /= GC_SUCCESS ) RETURN
-       State_Chm%IsorropBisulfate => NULL()
+       State_Chm%AteBisulfate => NULL()
     ENDIF
 
     IF ( ASSOCIATED( State_Chm%HPlusCloud ) ) THEN
@@ -3745,7 +3744,7 @@ CONTAINS
        IF ( RC /= GC_SUCCESS ) RETURN
        State_Chm%MnII_AMAX => NULL()
     ENDIF
-    
+
     IF ( ASSOCIATED( State_Chm%HSO3_AQ_A ) ) THEN
        DEALLOCATE( State_Chm%HSO3_AQ_A, STAT=RC )
        CALL GC_CheckVar( 'State_Chm%HSO3_AQ_A', 3, RC )
@@ -3759,7 +3758,7 @@ CONTAINS
        IF ( RC /= GC_SUCCESS ) RETURN
        State_Chm%SO3_AQ_A => NULL()
     ENDIF
-    
+
     IF ( ASSOCIATED( State_Chm%SO2_AQ_A ) ) THEN
        DEALLOCATE( State_Chm%SO2_AQ_A, STAT=RC )
        CALL GC_CheckVar( 'State_Chm%SO2_AQ_A', 3, RC )
@@ -4737,7 +4736,7 @@ CONTAINS
                                // ' from N2O5 fine aerosol uptake'
           IF ( isUnits ) Units = 'l'
           IF ( isRank  ) Rank = 3
-          
+
        CASE ( 'GAMMASO2TMI' )
           IF ( isDesc  ) Desc = 'Sticking coefficient for Gamma SO2+TMI/O2'
           IF ( isUnits ) Units = 'l'
@@ -4860,7 +4859,7 @@ CONTAINS
              'ISORROPIA H+ concentration, accumulation mode'
           IF ( isUnits ) Units = 'mol L-1'
           IF ( isRank  ) Rank  = 3
-          
+
        CASE( 'ISORROPIONIC' )
           IF ( isDesc  ) Desc  = &
              'ISORROPIA  IONIC strength'
@@ -4979,7 +4978,7 @@ CONTAINS
           IF ( isDesc  ) Desc  = 'Aerosol sulfite concentration'
           IF ( isUnits ) Units = 'mol L-1'
           IF ( isRank  ) Rank  =  3
- 
+
        CASE ( 'SO2_AQ_A' )
           IF ( isDesc  ) Desc  = 'Aerosol SO2aq concentration'
           IF ( isUnits ) Units = 'mol L-1'
@@ -5932,7 +5931,7 @@ CONTAINS
     modelId = N
     IF ( PerSpc == 'DRY' ) modelId = State_Chm%Map_DryDep(N)
     IF ( PerSpc == 'WET' ) modelId = State_Chm%Map_WetDep(N)
-    
+
     ! Point to the proper species, by modelId
     ThisSpc => State_Chm%SpcData(modelId)%Info